refactor(core,db): consolidate dependency-key and unbind policy
The Phase 1 guards landed as inline special cases scattered through the
generic store surface. Reify the two concepts they were expressing:

- interloper.asset.base.expected_dependency() is now the single reading
  of the bare/qualified dependency-key convention. Source._resolve_deps,
  DAG._check_requires, Store._wire_intra_deps and the relation write
  path all resolve through it instead of five hand-rolled parsers. The
  DAG contract check thereby compares resolved identities: a bare key
  now means the declarer's own source — not any upstream that happens
  to share the asset key.
- RelationDefinition.on_unbind ('detach' | 'block') expresses what
  explicitly unbinding a bound required slot does. Asset dependencies
  declare 'block'; the store's unbind guards derive from the vocabulary
  and the literal 'dependency' branches in the generic relation methods
  are gone.
- The relation read/write layer moves to store/relations.py
  (RelationMixin, which ComponentMixin now builds on): vocabulary
  resolution, slot validation, per-slot upsert (_upsert_relation), and
  the shared _blocked_unbinds helper replace the duplicated inline
  logic. components.py shrinks from 897 to 617 lines and keeps CRUD,
  children sync, deletion, and hydration only.

Tests mirror the split: relation-policy tests move to
tests/store/test_relations.py. No API surface changes.

By Digitl
