+++ packages/interloper-db/src/interloper_db/store/components.py
@@ -12,12 +12,8 @@ genuinely owns are applied where the row's ``kind`` demands them:
   status cascades through the parent's.
 - **job**: hydration drift-checks every target before reconstruction.
 
-Relation writes are validated against the row's relation vocabulary — the
-catalog class's definition is authoritative (a concrete class may extend its
-anchor's vocabulary; source-owned assets resolve through their parent), the
-kind's anchor is the drift fallback — including the declared slots and each
-slot's expected destination identity, and stamped with the denormalized
-``org_id``/``src_kind``/``dst_kind`` triple the composite foreign keys verify.
+Relation reads and writes live in the :class:`RelationMixin` layer this
+mixin builds on — see :mod:`interloper_db.store.relations`.
 """
 
 from __future__ import annotations
@@ -27,6 +23,7 @@ from typing import Any, cast
 from uuid import UUID
 
 import interloper as il
+from interloper.asset.base import expected_dependency
 from interloper.errors import (
     CatalogKeyError,
     ComponentDriftError,
@@ -35,17 +32,12 @@ from interloper.errors import (
     InUseError,
     NotFoundError,
 )
-from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import selectinload
 from sqlmodel import Session, col, select
 
 from interloper_db.drift import ComponentStatus, asset_status, resolve_component_cls, resolve_source_cls, source_status
 from interloper_db.models import Component, ComponentRelation
-from interloper_db.store.base import StoreBase
-
-# One relation binding: (destination component id, slot). Slot is "" for
-# slotless relation types.
-Binding = tuple[UUID, str]
+from interloper_db.store.relations import Binding, RelationMixin, _add_relation
 
 # Eager-load set for rows returned to API consumers: the parent, children
 # with their relations, and two hops (component → destination → resources).
@@ -60,8 +52,8 @@ COMPONENT_LOAD_OPTIONS = [
     .selectinload(ComponentRelation.dst),  # ty: ignore[invalid-argument-type]
 ]
 
-class ComponentMixin(StoreBase):
-    """Store methods for component CRUD, relations, and hydration."""
+class ComponentMixin(RelationMixin):
+    """Store methods for component CRUD and hydration, on the relation layer."""
 
     # -- CRUD ------------------------------------------------------------------
 
@@ -261,130 +253,6 @@ class ComponentMixin(StoreBase):
             {"id": str(c.id), "kind": c.kind, "key": c.key, "name": c.name}
             for c in sorted(referrers.values(), key=lambda c: ((c.name or c.key).lower(), str(c.id)))
         ]
-
-    def _relation_detaches(self, session: Session, db_src: Component, relation: ComponentRelation) -> bool:
-        """Whether a relation detaches (rather than blocks) when its destination is deleted.
-
-        Consults the referrer's own vocabulary: a type declared
-        ``on_delete="detach"`` detaches, as does a slot the referrer declares
-        optional (``RelationSlot.required=False`` — an ``optional_requires``
-        dependency). Anything unresolvable — unknown type, drifted key,
-        undeclared slot — blocks, keeping the guard fail-closed.
-        """
-        definition = self._relation_vocabulary(session, db_src).get(relation.type)
-        if definition is None:
-            return False
-        if definition.on_delete == "detach":
-            return True
-        slot = definition.slots.get(relation.slot)
-        return slot is not None and not slot.required
-
-    def _relation_vocabulary(self, session: Session, db_src: Component) -> dict[str, il.RelationDefinition]:
-        """A referrer row's relation vocabulary, slots included.
-
-        Source-owned assets are not top-level catalog entries: their concrete
-        definition (which carries the dependency slots' ``required`` flags)
-        comes from the parent source's definition. Everything else resolves
-        through the catalog with the kind's anchor as drift fallback.
-        """
-        if db_src.kind == "asset" and db_src.parent_id is not None:
-            parent = session.get(Component, db_src.parent_id)
-            parent_definition = self._catalog.get(parent.key) if parent else None
-            for asset_definition in getattr(parent_definition, "assets", []):
-                if asset_definition.key == db_src.key:
-                    return asset_definition.relations
-        return self._catalog.vocabulary(db_src.kind, db_src.key)
-
-    # -- Relations -------------------------------------------------------------
-
-    def list_relations(self, org_id: UUID, *, type: str | None = None) -> list[ComponentRelation]:
-        """List an organisation's component relations, optionally by type."""
-        with self._session() as session:
-            statement = select(ComponentRelation).where(ComponentRelation.org_id == org_id)
-            if type:
-                statement = statement.where(ComponentRelation.type == type)
-            return list(session.exec(statement).all())
-
-    def add_relation(self, component_id: UUID, *, type: str, dst_id: UUID, slot: str = "") -> ComponentRelation:
-        """Add one relation from a component.
-
-        Slotted types upsert per slot: re-binding an already-bound slot
-        repoints it to the new destination, and re-adding an identical
-        relation is a no-op returning the existing row.
-
-        Returns:
-            The relation (created, repointed, or already present).
-
-        Raises:
-            NotFoundError: If either endpoint is missing or cross-org.
-            ConfigError: If the kind's vocabulary doesn't declare the type,
-                the slot, or the destination doesn't match the slot's shape.
-        """
-        with self._session() as session:
-            src = session.get(Component, component_id)
-            if not src:
-                raise NotFoundError(f"Component {component_id} not found")
-            definition = self._check_vocabulary(session, src, type)
-            dst = self._resolve_dst(session, src, definition, type, slot, dst_id)
-            statement = select(ComponentRelation).where(
-                ComponentRelation.src_id == src.id,
-                ComponentRelation.type == type,
-                ComponentRelation.slot == slot,
-            )
-            if not definition.slotted:
-                statement = statement.where(ComponentRelation.dst_id == dst.id)
-            existing = session.exec(statement).all()
-            if match := next((relation for relation in existing if relation.dst_id == dst.id), None):
-                return match
-            for relation in existing:
-                session.delete(relation)
-            if existing:
-                session.flush()
-            relation = _add_relation(session, src, dst, type, slot)
-            try:
-                session.commit()
-            except IntegrityError:
-                raise ConfigError(
-                    f"Relation '{type}'{f' slot {slot!r}' if slot else ''} on '{src.key}' "
-                    f"was modified concurrently; retry"
-                )
-            return relation
-
-    def remove_relation(self, component_id: UUID, *, type: str, dst_id: UUID) -> None:
-        """Remove a component's relations of one type toward one destination.
-
-        Raises:
-            ConfigError: If a matching edge fills a *required* dependency
-                slot — unbinding it would break the asset at its next run;
-                repoint the slot or remove the dependent asset instead.
-        """
-        with self._session() as session:
-            statement = select(ComponentRelation).where(
-                ComponentRelation.src_id == component_id,
-                ComponentRelation.type == type,
-                ComponentRelation.dst_id == dst_id,
-            )
-            relations = session.exec(statement).all()
-            if type == "dependency" and relations:
-                src = session.get(Component, component_id)
-                definition = self._relation_vocabulary(session, src).get(type) if src else None
-                required = sorted(
-                    {
-                        relation.slot
-                        for relation in relations
-                        if definition is not None
-                        and (slot_def := definition.slots.get(relation.slot)) is not None
-                        and slot_def.required
-                    }
-                )
-                if required:
-                    raise ConfigError(
-                        f"Required dependency slot(s) {required} of '{src.key if src else component_id}' "
-                        f"cannot be unbound; repoint them or remove the dependent asset instead"
-                    )
-            for relation in relations:
-                session.delete(relation)
-            session.commit()
 
     # -- Hydration & status ----------------------------------------------------
 
@@ -678,143 +546,14 @@ class ComponentMixin(StoreBase):
             if child is None:
                 continue
             all_requires = {**asset_type.requires, **asset_type.optional_requires}
-            for param_name, required_qualified_key in all_requires.items():
-                # Only wire intra-source deps (qualified key belongs to this source)
-                if "." in required_qualified_key:
-                    qualified_source, qualified_asset = required_qualified_key.split(".", 1)
-                    if qualified_source != source_key:
-                        continue
-                else:
-                    qualified_asset = required_qualified_key
-                if qualified_asset == asset_key or qualified_asset not in children_by_key:
+            for param_name, declared_key in all_requires.items():
+                expected_source, expected_asset = expected_dependency(declared_key, own_source_key=source_key)
+                # Cross-source deps are never auto-wired; nor self-edges.
+                if expected_source != source_key or expected_asset == asset_key:
                     continue
-                if (child.id, param_name) in bound:
+                if expected_asset not in children_by_key or (child.id, param_name) in bound:
                     continue
-                _add_relation(session, child, children_by_key[qualified_asset], "dependency", param_name)
-
-    def _sync_relations(
-        self,
-        session: Session,
-        src: Component,
-        relations: dict[str, list[Binding]] | None,
-    ) -> None:
-        """Replace the relation types present in *relations* (empty list clears).
-
-        Required dependency slots that are currently bound must stay bound —
-        dropping one would break the asset at its next run. Repointing (same
-        slot, different destination) is allowed.
-        """
-        for type_, bindings in (relations or {}).items():
-            definition = self._check_vocabulary(session, src, type_)
-            existing = session.exec(
-                select(ComponentRelation).where(ComponentRelation.src_id == src.id, ComponentRelation.type == type_)
-            ).all()
-            if type_ == "dependency":
-                kept = {slot for _, slot in bindings}
-                dropped = sorted(
-                    {
-                        relation.slot
-                        for relation in existing
-                        if relation.slot not in kept
-                        and (slot_def := definition.slots.get(relation.slot)) is not None
-                        and slot_def.required
-                    }
-                )
-                if dropped:
-                    raise ConfigError(
-                        f"Required dependency slot(s) {dropped} of '{src.key}' cannot be unbound; "
-                        f"repoint them or remove the dependent asset instead"
-                    )
-            for relation in existing:
-                session.delete(relation)
-            session.flush()
-            for dst_id, slot in bindings:
-                dst = self._resolve_dst(session, src, definition, type_, slot, dst_id)
-                _add_relation(session, src, dst, type_, slot)
-
-    def _check_vocabulary(self, session: Session, src: Component, type_: str) -> il.RelationDefinition:
-        """Reject relation types the row's class vocabulary doesn't declare.
-
-        Returns:
-            The type's relation definition (slots included for owned assets).
-        """
-        vocabulary = self._relation_vocabulary(session, src)
-        if type_ not in vocabulary:
-            raise ConfigError(
-                f"Components of kind '{src.kind}' ('{src.key}') declare no '{type_}' relations "
-                f"(allowed: {sorted(vocabulary) or 'none'})"
-            )
-        return vocabulary[type_]
-
-    def _resolve_dst(
-        self, session: Session, src: Component, definition: il.RelationDefinition, type_: str, slot: str, dst_id: UUID
-    ) -> Component:
-        """Resolve a relation destination, enforcing existence, same-org, and the vocabulary's shape."""
-        dst = session.get(Component, dst_id)
-        if dst is None or dst.org_id != src.org_id:
-            raise NotFoundError(f"Component {dst_id} not found (relation '{type_}'{f'/{slot}' if slot else ''})")
-        if dst.id == src.id:
-            raise ConfigError(f"Relation '{type_}' on '{src.key}' may not point at the component itself")
-        if dst.kind not in definition.kinds:
-            raise ConfigError(
-                f"Relation '{type_}' on kind '{src.kind}' may not point at a '{dst.kind}' "
-                f"component (allowed: {definition.kinds})"
-            )
-        if not definition.slotted and slot:
-            raise ConfigError(f"Relation '{type_}' on kind '{src.kind}' is not slotted (got slot '{slot}')")
-        # A definition with no declared slots (a kind anchor reached through
-        # the drift fallback) can't vet slot names or targets — skip, fail-open.
-        if definition.slotted and definition.slots:
-            slot_def = definition.slots.get(slot)
-            if slot_def is None:
-                raise ConfigError(
-                    f"Relation '{type_}' on '{src.key}' declares no slot '{slot}' "
-                    f"(declared: {sorted(definition.slots)})"
-                )
-            self._check_slot_target(session, src, type_, slot, slot_def, dst)
-        return dst
-
-    def _check_slot_target(
-        self, session: Session, src: Component, type_: str, slot: str, slot_def: il.RelationSlot, dst: Component
-    ) -> None:
-        """Enforce a slot's declared destination identity, when it declares one.
-
-        Dependency slots carry an asset key — bare (a sibling of the same
-        source instance) or qualified ``source_key.asset_key``. A qualified
-        key naming the src's own source still binds within the instance;
-        any other source key accepts that source's assets from any instance.
-        Other slotted types (resources) declare a plain component key.
-        """
-        expected = slot_def.key
-        if not expected:
-            return
-        if type_ != "dependency":
-            if dst.key != expected:
-                raise ConfigError(
-                    f"Relation '{type_}' slot '{slot}' of '{src.key}' expects a '{expected}' "
-                    f"component, got '{dst.key}'"
-                )
-            return
-        expected_source, expected_asset = expected.split(".", 1) if "." in expected else ("", expected)
-        if dst.key != expected_asset:
-            raise ConfigError(
-                f"Dependency slot '{slot}' of '{src.key}' expects asset '{expected}', got '{dst.key}'"
-            )
-        src_parent = session.get(Component, src.parent_id) if src.parent_id else None
-        intra_source = not expected_source or (src_parent is not None and expected_source == src_parent.key)
-        if intra_source:
-            if src.parent_id is not None and dst.parent_id != src.parent_id:
-                raise ConfigError(
-                    f"Dependency slot '{slot}' of '{src.key}' must bind a sibling asset "
-                    f"of the same source instance"
-                )
-        else:
-            dst_parent = session.get(Component, dst.parent_id) if dst.parent_id else None
-            if dst_parent is None or dst_parent.key != expected_source:
-                raise ConfigError(
-                    f"Dependency slot '{slot}' of '{src.key}' expects an asset of source "
-                    f"'{expected_source}', got one of '{dst_parent.key if dst_parent else 'none'}'"
-                )
+                _add_relation(session, child, children_by_key[expected_asset], "dependency", param_name)
 
     def _row_status(self, session: Session, db_component: Component) -> ComponentStatus:
         """Row drift status inside an open session (parent fetched on demand)."""
@@ -876,22 +615,3 @@ def stamp_component_state(db_component: Component, **fields: Any) -> None:
     if model is not None:
         model.model_validate(state)
     db_component.state = state
-
-
-def _add_relation(session: Session, src: Component, dst: Component, type_: str, slot: str = "") -> ComponentRelation:
-    """Add one relation, stamping the denormalized org/kind triple from the rows.
-
-    Returns:
-        The pending relation (added to the session, not flushed).
-    """
-    relation = ComponentRelation(
-        src_id=src.id,
-        type=type_,
-        slot=slot,
-        dst_id=dst.id,
-        org_id=src.org_id,
-        src_kind=src.kind,
-        dst_kind=dst.kind,
-    )
-    session.add(relation)
-    return relation
